Lib Path, Decurry, Val Improuvement, Enum Converter Fix
Curry call
Val, var, let now detect function correctly
Enum now better parse CSV
Blocktags & Entitytags are now in subfolder according to path
Library path can be specified in Compiler Setting
Library moved for next version
Compiler can now be called from cmd

+++ JSharp/CMD_Compile.cs
@@ -220,14 +220,17 @@ namespace BluePhoenix
         }
         public void ExportResourcePack(string path)
         {
-            string rpPath = Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location) + "/tmp_rp";
-            SafeWriteFile(rpPath + "/pack.mcmeta",
-                            JsonConvert.SerializeObject(new ResourcePackMeta(project.projectName + " - " + project.description)));
-
-            if (Directory.Exists(rpPath))
+            if (path != null && path != "")
             {
-                if (File.Exists(path)) { File.Delete(path); }
-                ZipFile.CreateFromDirectory(rpPath, path);
+                string rpPath = Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location) + "/tmp_rp";
+                SafeWriteFile(rpPath + "/pack.mcmeta",
+                                JsonConvert.SerializeObject(new ResourcePackMeta(project.projectName + " - " + project.description)));
+
+                if (Directory.Exists(rpPath))
+                {
+                    if (File.Exists(path)) { File.Delete(path); }
+                    ZipFile.CreateFromDirectory(rpPath, path);
+                }
             }
         }
         public static void SafeCopy(string src, string dest)
